Rename Get operations for consistency
