+++ backend-api/src/functions/common/session/getOperations/GetSessionBiometricTokenIssued/GetSessionBiometricTokenIssued.test.ts
