+++ backend-api/src/functions/common/session/getOperations/GetSessionBiometricTokenIssued/GetSessionBiometricTokenIssued.ts
