+++ backend-api/src/functions/common/session/getOperations/GetSessionIssueBiometricCredential/GetSessionIssueBiometricCredential.test.ts
